refactor(frontend): move DatabaseChangeView to sidebar
Move the SQL checks and stages overview from a main content view to
the sidebar for consistent layout. The DatabaseChangeSection now
displays after issue-related sections (status, approval flow, labels).
